pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chatbot>=1.5.2b0",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "requests>=2.32.3",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
    "python-dotenv>=1.1.0",
]
